refactor(dashboard): relocate format functions from lib to lib/utils for better organization
