+++ src/app/dashboard/search/GroupCardSection.tsx
@@ -3,12 +3,12 @@ import { PhotoProvider, PhotoView } from 'react-photo-view';
 
 import { Button } from '@/components/ui/button';
 import { highlightText } from '@/components/ui/highlight-text';
-import { formatToShortDay } from '@/lib/format';
 import { getAccountLabelById } from '@/lib/utils/account';
 import {
   getCategoryIconById,
   getCategoryLabelById,
 } from '@/lib/utils/categories';
+import { formatToShortDay } from '@/lib/utils/format';
 import { Account } from '@/types/account';
 import { AccountingRecord } from '@/types/accounting';
 import { Category } from '@/types/category';
