+++ src/lib/utils/format.ts
@@ -16,7 +16,10 @@ export const formatToShortDay = (date: string | Date) => {
  * - 月視圖：yyyy-MM-dd
  * - 年視圖：MMM
  */
-export const formatChartTimeKey = (date: string | Date, isYearView: boolean) => {
+export const formatChartTimeKey = (
+  date: string | Date,
+  isYearView: boolean,
+) => {
   const formattedDate = new Date(date);
   return isYearView
     ? format(formattedDate, 'MMM')
